Refactor browser module
